% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_bod.R
\name{a_bod}
\alias{a_bod}
\title{Benefit of doubt aggregation}
\usage{
a_bod(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
As used in compind package. This is a wrapper which also returns `x` if `x`
only has one column (to avoid errors in Compind).
}
\examples{
#
}
