% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_wroclaw.R
\name{a_wroclaw}
\alias{a_wroclaw}
\title{Wroclaw Taxonomic aggregation}
\usage{
a_wroclaw(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
As used in Compind package. This is a wrapper which also returns `x` if `x`
only has one column (to avoid errors in Compind).
}
\examples{
#
}
