% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_add_indicators.R
\name{f_add_indicators}
\alias{f_add_indicators}
\title{Add indicators}
\usage{
f_add_indicators(coin, add_indicators = NULL)
}
\arguments{
\item{coin}{The coin}

\item{add_indicators}{Character vector of indicator codes to add back in.}
}
\value{
coin Updated coin.
}
\description{
As `f_remove_indicators()` but adds indicators back in. Obviously only
 indicators that were originally present in the input data can be added.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

# remove first
MVI <- f_remove_indicators(MVI, c("S.G.3", "A.M.1"))
# print method
MVI

# add one back
MVI <- f_add_indicators(MVI, "S.G.3")
MVI
}
