% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_analyse_indicators.R
\name{f_analyse_indicators}
\alias{f_analyse_indicators}
\title{Analyse indicators}
\usage{
f_analyse_indicators(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
coin Updated coin with analysis tables
}
\description{
Takes a coin and outputs an analysis of indicators. The objective is to look
for any "statistically-problematic" indicators, based on:
}
\details{
- data availability
- high proportion of repeated data values
- outliers as defined by skew and kurtosis
- collinearity within first aggregation level (category)
- negative correlation within first aggregation level (category)

The output can be viewed in R or as an interactive data frame.
Uses Spearman rank correlation to deal with skewed distributions.
Operates on the Raw data set.

The output is a coin with the results attached. This is so that on export, the
results will also be exported easily.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )


MVI <- f_analyse_indicators(MVI)
}
