% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_build_index.R
\name{f_build_index}
\alias{f_build_index}
\title{Build index}
\usage{
f_build_index(coin, agg_method = "a_amean")
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}

\item{agg_method}{One of `"a_amean"` (arithmetic mean), `"a_gmean"` (geometric mean),
`"a_bod"` (benefit of doubt via Compind package) or `"a_wroclaw"` (Wroclaw Taxonomic Method via Compind).}
}
\value{
coin COIN object
}
\description{
this function builds the MVI. Assumes that at this point you have imported
your data and built the MVI coin. Also optionally you have analysed and
possibly removed indicators, but taken no further steps.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))

f_build_index(coin = MVI)
}
