% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_change_weights.R
\name{f_change_weights}
\alias{f_change_weights}
\title{Change weights}
\usage{
f_change_weights(coin, w)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}

\item{w}{weight. The weights `w` can either be a named list with names as iCodes
          and values the new weights, OR 
          as a data frame with columns "iCode" and "Weight" with 
          codes and corresponding weights.   
          In both cases a subset of weight-code pairs can be specified.
         E.g. `list("Salud" = 0.5, Amenazas = 0.8)`.

         OR set w = "equal" for equal weights everywhere, or w = "original" to use the
          weights that were input with the input data.

          Remember that weights are relative within aggregation groups.}
}
\value{
coin COIN object
}
\description{
Function that takes some new weights and regenerates the coin.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))
MVI2 <- f_analyse_indicators(coin = MVI)
MVI3 <- f_build_index(coin = MVI2)
MVI4 <- f_change_weights(coin = MVI3, w= list(Amenazas = 2, Cap_Resp = 0.5))
MVI4
}
