% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_data_input.R
\name{f_data_input}
\alias{f_data_input}
\title{Data input}
\usage{
f_data_input(file_path)
}
\arguments{
\item{file_path}{path to the excel file where we have the raw data - organised
with the format exepected by COINR package}
}
\value{
coin object with a COINR class
}
\description{
Outputs a constructed coin
Reads an Excel file found at `file_path` which is expected to be in a specific
format. See "inst/data_input/data_module-input.xlsx" for example.
.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))

}
