% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_display_indicator_analysis.R
\name{f_display_indicator_analysis}
\alias{f_display_indicator_analysis}
\title{Display indicator analysis}
\usage{
f_display_indicator_analysis(coin, filter_to_flagged = TRUE)
}
\arguments{
\item{coin}{The coin.}

\item{filter_to_flagged}{Logical: if `TRUE` filters to only show indicators with
at least one flag.}
}
\value{
An interactive table (DT object).
}
\description{
Displays the analysis performed by [f_analyse_indicators()] in an interactive
table using the DT package.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

MVI <- f_analyse_indicators(MVI)

f_display_indicator_analysis(MVI)
}
