% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_display_results_table.R
\name{f_display_results_table}
\alias{f_display_results_table}
\title{Display results table}
\usage{
f_display_results_table(coin, type = "scores")
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}

\item{type}{set type = "scores" or "ranks"}
}
\value{
a visualised data table
}
\description{
Outputs an interactive results table suitable for HTML documents and the app.
.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))

MVI2 <- f_build_index(coin = MVI)
f_display_results_table(coin = MVI2)
}
