% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_export_to_excel.R
\name{f_export_to_excel}
\alias{f_export_to_excel}
\title{Export to excel}
\usage{
f_export_to_excel(coin, fname = "index_export.xlsx")
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}

\item{fname}{file name}
}
\description{
Simplified export to Excel with some formatting
Note if we change the index structure this may need adjusting.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))
MVI2 <- f_analyse_indicators(coin = MVI)
MVI3 <- f_build_index(coin = MVI2)

f_export_to_excel(coin = MVI3, 
                  fname = here::here("inst", "index_export.xlsx"))
}
