% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_gather_correlations.R
\name{f_gather_correlations}
\alias{f_gather_correlations}
\title{Gather correlations}
\usage{
f_gather_correlations(X)
}
\description{
Helper function for indicator analysis
}
\examples{
#f_gather_correlations()
}
