% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_generate_results.R
\name{f_generate_results}
\alias{f_generate_results}
\title{Generate results}
\usage{
f_generate_results(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
Updated coin.
}
\description{
Generates sorted results tables and attaches back to the coin. This is a helper
function which is called by [f_build_index()].
}
\examples{
#
}
