% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_equal_weights.R
\name{f_get_equal_weights}
\alias{f_get_equal_weights}
\title{Get equal weights}
\usage{
f_get_equal_weights(coin)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}
}
\value{
w weight
}
\description{
Description
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))
MVI2 <- f_analyse_indicators(coin = MVI)
MVI3 <- f_remove_indicators(coin = MVI2, c("S.G.3", "A.M.1"))
MVI4 <- f_build_index(coin = MVI3)
weight <- f_get_equal_weights(coin = MVI4)
weight
}
