% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_last_weights.R
\name{f_get_last_weights}
\alias{f_get_last_weights}
\title{Get last weights}
\usage{
f_get_last_weights(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
Data frame of weights.
}
\description{
Returns the data frame of weights last used to aggregate the coin. Helper
called by [f_change_weights()].
}
\examples{
#
}
