% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_highlight_dt.R
\name{f_highlight_DT}
\alias{f_highlight_DT}
\title{Highlight Data Table}
\usage{
f_highlight_DT(Xd, Xh, table_caption = NULL, highlight_colour = "#ffc266")
}
\value{
A DT object
}
\description{
Generic function for creating an interactive table with input df `Xd`, and
with cells highlighted by an equivalently sized logical df `Xh`.
This is used for displaying the output of `f_analyse_indicators()`.
}
\examples{
#f_highlight_DT()
}
