% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_framework.R
\name{f_plot_framework}
\alias{f_plot_framework}
\title{Interactive sunburst plot of index structure}
\usage{
f_plot_framework(coin, seg_cols = NULL)
}
\arguments{
\item{seg_cols}{A character vector of colour codes, one for each segment in the plot. The length of this
vector must be equal to the number of segments, i.e. the sum of the number of indicators and aggregates
in each level.}

\item{COIN}{coin object}
}
\value{
fig as a plotly plot
}
\description{
Plots the structure of the index using a sunburst plot using **plotly**.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

f_plot_framework(MVI)
}
