% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_print_coin.R
\name{f_print_coin}
\alias{f_print_coin}
\title{Print Coin}
\usage{
f_print_coin(coin)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}

\item{seg_cols}{A character vector of colour codes, one for each segment in 
the plot. The length of this #' vector must be equal to the number
 of segments, i.e. the sum of the number of indicators and 
 aggregates in each level.}
}
\description{
Print Coin
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))
f_print_coin(MVI)
}
