% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_remove_indicators.R
\name{f_remove_indicators}
\alias{f_remove_indicators}
\title{Remove indicators}
\usage{
f_remove_indicators(coin, remove_indicators = NULL)
}
\arguments{
\item{coin}{A coin.}

\item{remove_indicators}{Character vector of indicator codes to remove.}
}
\value{
coin COIN object
}
\description{
Removes indicators as specified by a character vector of indicator
codes in `remove_indicators`. After removal, any results are regenerated
(updated).
}
\details{
The original data is always preserved so indicators can be restored.
Analysis tables are not currently re-run so this would have to be done separately.
However, there is actually no need to re-run.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

# call print method
MVI

MVI <- f_remove_indicators(MVI, c("S.G.3", "A.M.1"))
# compare with previous
MVI
}
