% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solo_groups.R
\name{get_solo_groups}
\alias{get_solo_groups}
\title{Codes of any groups with only one child}
\usage{
get_solo_groups(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
a coin
}
\description{
To avoid errors in Compind functions
}
\examples{
#
}
