% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_treatment_details.R
\name{tidy_treatment_details}
\alias{tidy_treatment_details}
\title{Data Treatment Pipeline}
\usage{
tidy_treatment_details(X)
}
\arguments{
\item{X}{corresponding to a specific part of the COIN 
MVI$Analysis$Treated$Dets_Table}
}
\value{
a summary table
}
\description{
Tidy the results of data treatment
}
\examples{
#tidy_treatment_details()
}
