% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_add_indicators.R
\name{f_add_indicators}
\alias{f_add_indicators}
\title{Add indicators}
\usage{
f_add_indicators(coin, add_indicators = NULL)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}

\item{add_indicators}{}
}
\value{
coin COIN object
}
\description{
As `f_remove_indicators()` but adds indicators back in. Obviously only
 indicators that were originally present in the input data can be added.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))

f_add_indicators(MVI)
}
