% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_analyse_indicators.R
\name{f_analyse_indicators}
\alias{f_analyse_indicators}
\title{Analyse indicators}
\usage{
f_analyse_indicators(coin)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}
}
\value{
coin COIN object
}
\description{
Takes a coin and outputs an analysis of indicators. The objective is to look
for any "statistically-problematic" indicators, based on:
- data availability
- high proportion of repeated data values
- outliers as defined by skew and kurtosis
- collinearity within first aggregation level (category)
- negative correlation within first aggregation level (category)
}
\details{
The output can be viewed in R or as an interactive data frame.
Uses Spearman rank correlation to deal with skewed distributions.
Operates on the Raw data set.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))
f_analyse_indicators(MVI)
}
