% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_build_index.R
\name{f_build_index}
\alias{f_build_index}
\title{Build index}
\usage{
f_build_index(coin, agg_method = "a_amean")
}
\arguments{
\item{coin}{The coin.}

\item{agg_method}{One of `"a_amean"` (arithmetic mean), `"a_gmean"` (geometric mean),
`"a_bod"` (benefit of doubt via Compind package) or `"a_wroclaw"` (Wroclaw Taxonomic Method via Compind).}
}
\value{
coin Updated coin.
}
\description{
This function builds the index. It assumes that at this point you have imported
your data and built the coin. Also optionally you have analysed and
possibly removed indicators, but taken no further steps.
}
\details{
The operations performed here are:

- Treatment of outliers using a standard Winsorisation/log approach
- Normalistion to the 1, 100 interval
- Aggregation using one of the methods specified by `agg_method`

The output is an updated coin.

NOTE that for the Compind-powered aggregation methods, we need to be careful
about how the aggregated values are rescaled or not in the case of a single
indicator in an aggregation group. I initially blocked single-indicator groups
here but relaxed this after adjusting the aggregation wrapper functions.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

MVI <- f_build_index(coin = MVI)
}
