% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_change_weights.R
\name{f_change_weights}
\alias{f_change_weights}
\title{Change weights}
\usage{
f_change_weights(coin, w)
}
\arguments{
\item{coin}{The coin.}

\item{w}{weight. The weights `w` can either be a named list with names as iCodes
          and values the new weights, OR 
          as a data frame with columns "iCode" and "Weight" with 
          codes and corresponding weights.   
          In both cases a subset of weight-code pairs can be specified.
         E.g. `list("Salud" = 0.5, Amenazas = 0.8)`.

         OR set `w = "equal"` for equal weights everywhere, or `w = "original"` to use the
          weights that were input with the input data.

          Remember that weights are relative within aggregation groups.}
}
\value{
Updated coin.
}
\description{
Function that takes some new weights and regenerates the coin. Results are
regenerated from the aggregation step only, so all previously-specified methodology
is fixed.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

MVI <- f_build_index(MVI)
MVI2 <- f_change_weights(MVI, w= list(Amenazas = 1.5, Cap_Resp = 0.5))

# we can use a COINr function for a comparison
COINr::compare_coins(MVI, MVI2, dset = "Aggregated", iCode = "MVI") |>
  head()
}
