% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_data_input.R
\name{f_data_input}
\alias{f_data_input}
\title{Data input}
\usage{
f_data_input(file_path)
}
\arguments{
\item{file_path}{path to the excel file where we have the raw data - organised
with the format expected by COINr package}
}
\value{
coin-class object
}
\description{
Reads a formatted Excel file found at `file_path` and outputs a constructed coin.
}
\details{
On reading the Excel file, this function does the following:

- Data is split into data and metadata and tidied
- Metadata is merged with hard-coded index structure
- Any indicators with no data at all are removed
- Any resulting aggregation groups with no "children" are removed
- A coin is assembled using COINr and this is the function output

If indicators/groups are removed, a message is sent to the console.

The Excel file is required to be in a fairly strict format: an example is given at
`inst/data_input/data_module-input.xlsx`. This template is still a work in progress
and can be modified in the app phase following further feedback. See also comments
in the main vignette.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

}
