% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_display_indicator_analysis.R
\name{f_display_indicator_analysis}
\alias{f_display_indicator_analysis}
\title{Display indicator analysis}
\usage{
f_display_indicator_analysis(coin, filter_to_flagged = TRUE)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}

\item{filter_to_flagged}{TRUE or FALSE}
}
\description{
Description
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))

MVI2 <- f_analyse_indicators(MVI)

f_display_indicator_analysis(MVI2)
}
