% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_display_results_table.R
\name{f_display_results_table}
\alias{f_display_results_table}
\title{Display results table}
\usage{
f_display_results_table(coin, type = "scores")
}
\arguments{
\item{coin}{The coin.}

\item{type}{set One of `"scores"` or `"ranks"`.}
}
\value{
An interactive table using DT.
}
\description{
Outputs an interactive results table suitable for HTML documents and the app.
Can be further formatted (e.g. conditional formatting) to preference in the 
app stage. This is just a demo.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

MVI <- f_build_index(MVI)
f_display_results_table(MVI)
}
