% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_export_to_excel.R
\name{f_export_to_excel}
\alias{f_export_to_excel}
\title{Export to excel}
\usage{
f_export_to_excel(coin, fname = "index_export.xlsx")
}
\arguments{
\item{coin}{The coin, with results present.}

\item{fname}{file name to write the results to.}
}
\value{
Writes an Excel spreadsheet.
}
\description{
Simplified export to Excel with some formatting
Note if we change the index structure, the conditional formatting will need
to be adjusted.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))
MVI <- f_analyse_indicators(MVI)
MVI <- f_build_index(MVI)

f_export_to_excel(coin = MVI, 
                  fname = here::here("inst", "index_export.xlsx"))
}
