% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_gather_correlations.R
\name{f_gather_correlations}
\alias{f_gather_correlations}
\title{Gather correlations}
\usage{
f_gather_correlations(X)
}
\arguments{
\item{X}{a frame with the 2 indicators to check `"Ind1"`  Ind1 &  `"Ind1"`}
}
\value{
nothing a report is created...
}
\description{
Helper function for indicator analysis
}
\examples{
#f_gather_correlations()
}
