% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_generate_results.R
\name{f_generate_results}
\alias{f_generate_results}
\title{Generate results}
\usage{
f_generate_results(coin)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}
}
\value{
coin COIN object
}
\description{
Generates sorted results tables and attaches back to the coin
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))

MVI2 <- f_build_index(coin = MVI)

f_generate_results(coin = MVI2)
}
