% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_equal_weights.R
\name{f_get_equal_weights}
\alias{f_get_equal_weights}
\title{Get equal weights}
\usage{
f_get_equal_weights(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
A data frame of weights.
}
\description{
Returns a weight data frame of equal weights. Helper called by [f_change_weights()].
}
\examples{
#
}
