% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_get_last_weights.R
\name{f_get_last_weights}
\alias{f_get_last_weights}
\title{Get last weights}
\usage{
f_get_last_weights(coin)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}
}
\value{
w weight
}
\description{
Description
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))
MVI2 <- f_analyse_indicators(MVI)
MVI3 <- f_build_index(coin = MVI2)
weight <- f_get_last_weights(MVI3)
weight 
}
