% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_map.R
\name{f_plot_map}
\alias{f_plot_map}
\title{Plot map}
\usage{
f_plot_map(coin, dset = "Aggregated", iCode = "MVI", shp_path)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}

\item{shp_path}{is currently at "shp/gtm_admbnda_adm2_ocha_conred_20190207.shp"}
}
\description{
Plots an interactive choropleth map of the index or any indicator, using 
supplied shape files.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))

MVI2 <- f_build_index(coin = MVI)

f_plot_map(coin = MVI2, 
           dset = "Aggregated",
           iCode = "MVI", 
           shp_path = system.file("shp/gtm_admbnda_adm2_ocha_conred_20190207.shp",
                                            package = "BuildIndex")  )
## when using ut your shape in the folder ---
         #  shp_path = here::here("inst/shp", "gtm_admbnda_adm2_ocha_conred_20190207.shp")           )
}
