% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_map.R
\name{f_plot_map}
\alias{f_plot_map}
\title{Plot map}
\usage{
f_plot_map(coin, dset = "Aggregated", iCode = "MVI", shp_path)
}
\arguments{
\item{coin}{The coin}

\item{dset}{Name of data set in the coin from which to extract the indicator/aggregate to plot}

\item{iCode}{Code of indicator within `dset` to plot}

\item{shp_path}{Assume a shapefile with ADM2_PCODE being the main key to use\
with thedefault admin2Pcode within the indicator main excel file

 Guatemala example is currently at `"./inst/shp/gtm_admbnda_adm2_ocha_conred_20190207.shp"`}
}
\value{
a leaflet object
}
\description{
Plots an interactive choropleth map of the index or any indicator, using 
supplied shape files. This is powered by the leaflet package.
}
\details{
Note we will need some checks here to make sure the shape file is compatible with the
data. This can be dealt with in the app phase.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

shp_path <-  system.file("data_input/gtm_admbnda_adm2_ocha_conred_20190207.shp",
                                            package = "BuildIndex")


MVI <- f_build_index(MVI)

f_plot_map(coin = MVI, 
           dset = "Aggregated",
           iCode = "MVI", 
           shp_path = shp_path  )
## when using ut your shape in the folder ---
         #  shp_path = here::here("inst/data-input", "gtm_admbnda_adm2_ocha_conred_20190207.shp")           )
}
