% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plot_map2.R
\name{f_plot_map2}
\alias{f_plot_map2}
\title{Plot map with main index and sub index}
\usage{
f_plot_map2(
  coin,
  dset = "Aggregated",
  shp_path,
  shp_name,
  shp_ucode,
  dTolerance = 1000,
  palette = "inferno"
)
}
\arguments{
\item{coin}{The coin}

\item{dset}{Name of data set in the coin from which to extract the 
indicator/aggregate to plot}

\item{shp_path}{Assume a shapefile matching  with ucode
Guatemala example is currently at
 `"./inst/shp/gtm_admbnda_adm2_ocha_conred_20190207.shp"`}

\item{shp_name}{name of corresponding admin unit in the shapefile "ADM2_ES"}

\item{shp_ucode}{name of corresponding admincode matching  with ucode 
in the shapefile  for instance `"ADM2_PCODE"` being the main key to use}

\item{dTolerance}{1000 per default - parameter to simplify the shapefile so that the report is not too heavy...}

\item{palette}{color palette for the map - selected as "inferno" per default to be inlusive of the 10% population of color blind people}
}
\value{
a leaflet object
}
\description{
This second version of the summary map display the main severity index 
 together with all its sub-dimensions each them presented as selected layers
 
The infobox present the index (or subindex) value together with the rank
}
\details{
Plots an interactive choropleth map using supplied shape files. The shapefile is 
automatically simplified so that the resulting html viz remains light

This is powered by the leaflet package.

Note we will need some checks here to make sure the shape file is compatible with the
data. This can be dealt with in the app phase.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

MVI <- f_build_index(MVI)

shp_path <-  system.file("data_input/gtm_admbnda_adm2_ocha_conred_20190207.shp",
                                            package = "BuildIndex")

shp_name <- "ADM2_ES"
shp_ucode <- "ADM2_PCODE"

f_plot_map2(coin = MVI, 
           dset = "Aggregated",
           shp_path = shp_path,
           shp_name = shp_name,
           shp_ucode = shp_ucode )
## when using ut your shape in the folder ---
         #  shp_path = here::here("inst/data-input", "gtm_admbnda_adm2_ocha_conred_20190207.shp")           )
}
