% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_print_coin.R
\name{f_print_coin}
\alias{f_print_coin}
\title{Print Coin}
\usage{
f_print_coin(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
A coin
}
\description{
This is a print-style text output function for summarising the contents of the coin.
}
\details{
It is intended to be used immediately on loading data, to show the user what they have
input, so they can check for any unexpected things.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

f_print_coin(MVI)
}
