% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_remove_indicators.R
\name{f_remove_indicators}
\alias{f_remove_indicators}
\title{Remove indicators}
\usage{
f_remove_indicators(coin, remove_indicators = NULL)
}
\arguments{
\item{coin}{COIN object, or list with first entry is the indicator metadata, 
second entry is the aggregation metadata}

\item{remove_indicators}{}
}
\value{
coin COIN object
}
\description{
Removes indicators as specified by a character vector of indicator
codes in `remove_indicators`. After removal, any results are regenerated
(updated).
}
\details{
The original data is always preserved so indicators can be restored.
Analysis tables are not currently re-run so this has to be done separately.
}
\examples{
MVI <- f_data_input(file_path = system.file("data_input/data_module-input.xlsx",
                                            package = "BuildIndex") )

## when using create a data-raw folder and put you data input xlsx file there
# MVI <- f_data_input(here::here("data-raw", "data_module-input.xlsx"))
MVI2 <- f_analyse_indicators(MVI)
MVI3 <- f_remove_indicators(MVI2, c("S.G.3", "A.M.1"))
MVI3
}
